Move internal api file into src and make it a ts file
