+++ extension.bundle.ts
@@ -32,6 +32,7 @@ export * from './src/services/AzureResourcesService';
 // Export activate/deactivate for main.js
 export { activate, deactivate } from './src/extension';
 export * from './src/extensionVariables';
+export * from './src/hostapi.v2.internal';
 export * from './src/tree/BranchDataItemWrapper';
 export * from './src/tree/InvalidItem';
 export { ResourceGroupsItem } from './src/tree/ResourceGroupsItem';
